requirements.txt
allure-behave==2.14.2
allure-python-commons==2.14.2
Appium-Python-Client==5.1.1
attrs==25.3.0
behave==1.2.6
certifi==2025.4.26
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
parse==1.20.2
parse_type==0.6.4
pluggy==1.6.0
PySocks==1.7.1
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
